реорганизаия файлов проетка, изменение форм регистрации и логина
