--- text.csv
@@ -1,3 +1,0 @@
-Дата формирования;Владелец;Целевые взносы;Членские взносы
-05.12.2023;владелец уч 11;8000;2250
-05.12.2023;владелец уч 12;-3000;0
