+++ webapp/templates/registration.html
@@ -1,0 +1,50 @@
+<!doctype html>
+<html lang="ru" data-bs-theme="auto" xmlns="http://www.w3.org/1999/html">
+<head>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
+            integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
+            crossorigin="anonymous"></script>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="description" content="">
+    <title>Регистрация</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
+          integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
+    <link rel="stylesheet" href="{{ url_for('static', filename='sign-in.css') }}">
+</head>
+<body class="d-flex align-items-center py-4 bg-body-tertiary">
+<main class="form-signin w-100 m-auto">
+    <form>
+        <h1 class="h3 mb-3 fw-normal text-center">Зарегистрируйтесь</h1>
+        <div class="form-floating">
+            <input type="area" class="form-control" id="floatingInput" placeholder="12">
+            <label for="floatingInput">Введите номер участка</label>
+        </div>
+        <div class="form-floating">
+            <input type="email" class="form-control" id="floatingInput" placeholder="name@example.com">
+            <label for="floatingInput">Введите e-mail</label>
+        </div>
+
+        <div class="form-floating">
+            <input type="text" class="form-control" id="floatingInput" placeholder="+79262521235">
+            <label for="floatingInput">Введите номер телефона</label>
+
+        </div>
+        <div class="form-floating">
+            <input type="password1" class="form-control" id="floatingPassword" placeholder="Password">
+            <label for="floatingPassword">Введите пароль</label>
+        </div>
+        <div class="form-floating">
+            <input type="password" class="form-control" id="floatingPassword" placeholder="Password">
+            <label for="floatingPassword">Введите пароль еще раз</label>
+        </div>
+        <button class="btn btn-primary w-100 py-2" type="submit">Зарегистрироваться</button>
+        <p class="mt-2 mb-3 text-body-secondary">&copy;2023</p>
+    </form>
+</main>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
+        crossorigin="anonymous"></script>
+
+</body>
+</html>
